fix: 하단의 페이지 이동 버튼의 높이가 전송되지 않아 잘리는 현상 해결 페이지 이동 버튼이 늦게 로드됨에 따라 iframe의 height에 반영이 안됨 -> 페이지 이동 버튼이 잘려서 부모 서비스에서 보이지 않음
<해결>
기존 리스트가 업데이트 될 때마다 height를 갱신하도록 했던 방법에서
-> 리스트를 감싸고 있는 container의 높이에 변화가 생기는 걸 감지하여 height 갱신하는 방식으로 변경
+ height 전송 로직이 복잡해짐에 따라 message 전송 hook과 분리
